Don't install eslint globally
Prefer to install inside node_modules of each project.

A [RFC] for eslint proposed simplifying the process of package loading.
This change was implemented in eslint 6.0.0 and [reduced the level of
support] for
installing plugins globally

A ["highly dangerous and ugly package"] was published as a workaround.

At the same time eslint addressed [another issue] about loading
packages.

After [this change] packages are loaded relative to where they are specified in
a configuration file.

One simple way for the configuration and the location of the
configuration files to match is to install from the same `package.json`.

[another issue]: https://github.com/eslint/eslint/issues/10125
["highly dangerous and ugly package"]:
  https://github.com/wintercounter/eslint-global-patch
[rfc]: https://github.com/eslint/rfcs/pull/7
[reduced the level of support]:
  https://eslint.org/docs/user-guide/migrating-to-6.0.0#plugins-and-shareable-configs-are-no-longer-affected-by-eslints
[this change]:
  https://github.com/eslint/eslint/commit/25cc63d47e6c0aea8b88589a088c1f0de5f6f1cc